\name{Lipid}
\docType{data}
\alias{Lipid}
\title{Lipid and diabetes}
\description{
  This data set gives the simulated data for lipid, age, gender, and diabetes.
}
\usage{Lipid}
\format{A data frame containing 2000 observations.}
\source{simulated}
\references{Not applicable}
\keyword{datasets}