% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlor.r
\name{nlor}
\alias{nlor}
\title{Odds ratio plot for dose - response non-linear continuous exposure.}
\usage{
nlor(outcome, exposure, covar = NULL, ref = NULL, knum = 4, data)
}
\arguments{
\item{outcome}{the outcome variable}

\item{exposure}{the exposure variable}

\item{covar}{a covariats list}

\item{ref}{reference value for the continuous variable}

\item{knum}{number of knots}

\item{data}{name of a dataset}
}
\description{
Calculates non-linear odds ratio and plot OR vs. a continuous variable.
}
\examples{
sum1 <- nlor('dm', 'lipid', covar = c('age', 'gender'), 0.6, data = Lipid)
head(sum1)
}

