% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlorplot.r
\name{nlorplot}
\alias{nlorplot}
\title{Odds ratio plot for dose - response non-linear continuous exposure.}
\usage{
nlorplot(exposure, or, data, xlab = NULL)
}
\arguments{
\item{exposure}{the exposure variable}

\item{or}{odds ratio}

\item{data}{name of a dataset}

\item{xlab}{x-axis}
}
\description{
Calculates non-linear odds ratio and plot OR vs. a continuous variable.
}
\examples{
sum1 <- nlor('dm', 'lipid', covar = c('age', 'gender'), 0.6, data = Lipid)
head(sum1)
nlorplot('lipid', 'or', data = sum1, xlab = 'Lipid')
}

